irc: cleaner code
