+++ src/plugins/ircColors/index.ts
@@ -20,7 +20,7 @@ import { definePluginSettings } from "@api/Settings";
 import { hash as h64 } from "@intrnl/xxhash64";
 import { Devs } from "@utils/constants";
 import definePlugin, { OptionType } from "@utils/types";
-import { ChannelStore } from "@webpack/common";
+import { ChannelStore, GuildMemberStore, SelectedChannelStore } from "@webpack/common";
 
 function calculateHSLforId(id: string) {
     // No hooks here because it breaks RoleColorsEverywhere
@@ -32,6 +32,18 @@ function calculateHSLforId(id: string) {
         saturation: 100,
         lightness: settings.store.lightness
     };
+}
+
+// https://stackoverflow.com/a/44134328
+function hslToHex(hue, saturation, lightness) {
+    lightness /= 100;
+    const a = saturation * Math.min(lightness, 1 - lightness) / 100;
+    const f = n => {
+        const k = (n + hue / 30) % 12;
+        const color = lightness - a * Math.max(Math.min(k - 3, 9 - k, 1), -1);
+        return Math.round(255 * color).toString(16).padStart(2, "0"); // convert to Hex and prefix "0" if needed
+    };
+    return `#${f(0)}${f(8)}${f(4)}`;
 }
 
 // Calculate a CSS color string based on the user ID
@@ -93,31 +105,26 @@ export default definePlugin({
     ],
     // Propped to be used in TypingTweaks/RoleColorsEverywhere
     calculateHSLforId,
-    calculateNameColorForUser,
+    resolveUsedColor(userId: string) {
+        const channelId = SelectedChannelStore.getChannelId();
+        const channel = ChannelStore.getChannel(channelId);
+        const colorString = GuildMemberStore.getMember(channel.guild_id, userId)?.colorString;
+
+        if (settings.store.applyColorOnlyInDms && !channel?.isPrivate()) return colorString;
+        if (settings.store.applyColorOnlyToUsersWithoutColor && colorString) return colorString;
+
+        const { hue, lightness, saturation } = calculateHSLforId(userId);
+
+        return hslToHex(hue, saturation, lightness);
+    },
     calculateNameColorForMessageContext(message: any, colorString: string) {
         const id = message?.author?.id;
-        const color = calculateNameColorForUser(id);
-        const channel = ChannelStore.getChannel(message.channel_id);
 
-        if (settings.store.applyColorOnlyInDms && !channel?.isPrivate()) {
-            return colorString;
-        }
+        return id ? this.resolveUsedColor(id) : undefined;
 
-        return (!settings.store.applyColorOnlyToUsersWithoutColor || !colorString)
-            ? color
-            : colorString;
     },
     calculateNameColorForListContext(context: any) {
         const id = context?.user?.id;
-        const colorString = context?.colorString;
-        const color = calculateNameColorForUser(id);
-
-        if (settings.store.applyColorOnlyInDms && !context?.channel?.isPrivate()) {
-            return colorString;
-        }
-
-        return (!settings.store.applyColorOnlyToUsersWithoutColor || !colorString)
-            ? color
-            : colorString;
+        return id ? this.resolveUsedColor(id) : undefined;
     }
 });
