+++ src/plugins/roleColorEverywhere/index.tsx
@@ -72,20 +72,6 @@ const settings = definePluginSettings({
     }
 });
 
-// Needed to convert colors from IrcColors
-// https://stackoverflow.com/a/44134328
-function hslToHex(hue, saturation, lightness) {
-    lightness /= 100;
-    const a = saturation * Math.min(lightness, 1 - lightness) / 100;
-    const f = n => {
-        const k = (n + hue / 30) % 12;
-        const color = lightness - a * Math.max(Math.min(k - 3, 9 - k, 1), -1);
-        return Math.round(255 * color).toString(16).padStart(2, "0"); // convert to Hex and prefix "0" if needed
-    };
-    return `#${f(0)}${f(8)}${f(4)}`;
-}
-
-
 export default definePlugin({
     name: "RoleColorEverywhere",
     authors: [Devs.KingFish, Devs.lewisakura, Devs.AutumnVN, Devs.Kyuuhachi, Devs.jamesbt365],
@@ -178,10 +164,8 @@ export default definePlugin({
     getColorString(userId: string, channelOrGuildId: string) {
         if (Vencord.Plugins.isPluginEnabled("IrcColors")) {
             // @ts-ignore
-            const colors = Vencord.Plugins.plugins.IrcColors?.calculateHSLforId?.(userId);
-            if (colors) {
-                return hslToHex(colors.hue, colors.saturation, colors.lightness);
-            }
+            const color = Vencord.Plugins.plugins.IrcColors?.resolveUsedColor?.(userId);
+            if (color) return color;
         }
 
         try {
