+++ src/plugins/typingTweaks/index.tsx
@@ -58,13 +58,13 @@ interface Props {
 }
 
 const getUserColor = (userId: string, guildId: string) => {
+    const roleColor = GuildMemberStore.getMember(guildId, userId)?.colorString;
+
     if (Vencord.Plugins.isPluginEnabled("IrcColors")) {
         // @ts-ignore
-        const color = Vencord.Plugins.plugins?.IrcColors?.calculateNameColorForUser?.(userId);
+        const color = Vencord.Plugins.plugins.IrcColors?.resolveUsedColor?.(userId);
         if (color) return color;
     }
-
-    return GuildMemberStore.getMember(guildId, userId)?.colorString;
 };
 const TypingUser = ErrorBoundary.wrap(function ({ user, guildId }: Props) {
     return (
